+++ packages/tokens/src/fonts.ts
@@ -3,11 +3,3 @@ export const fonts = {
 	code: 'monospace',
 	'sf-digital': 'Lato, sans-serif',
 }
-
-export const mobileFonts = {
-	'thin-100': 'thin-100',
-	'light-300': 'light-300',
-	'regular-400': 'regular-400',
-	'bold-700': 'bold-700',
-	'black-900': 'black-900',
-}
