+++ packages/tokens/src/mobile-fonts.ts
@@ -1,0 +1,7 @@
+export const mobileFonts = {
+	'thin-100': 'thin-100',
+	'light-300': 'light-300',
+	'regular-400': 'regular-400',
+	'bold-700': 'bold-700',
+	'black-900': 'black-900',
+}
